target-ppc: fix mullw/mullwo
Based on patch by Julian Seward.


git-svn-id: svn://svn.savannah.nongnu.org/qemu/trunk@5379 c046a42c-6fe2-441c-8c8c-71466251a162

+++ target-ppc/op_helper.c
@@ -227,7 +227,7 @@ void do_divduo (void)
 
 void do_mullwo (void)
 {
-    int64_t res = (int64_t)T0 * (int64_t)T1;
+    int64_t res = (int64_t)(int32_t)T0 * (int64_t)(int32_t)T1;
 
     if (likely((int32_t)res == res)) {
         xer_ov = 0;
